product list improvements. added more fields; Created Component to handle the Filters
