+++ frontend/src/components/ProductList.css
@@ -40,7 +40,7 @@
 }
 
 .delete-button {
-  background-color: #ff4d4d;
+  background-color: #dc3545;
   color: white;
   border: none;
   border-radius: 5px;
@@ -56,4 +56,39 @@
 
 .delete-button {
   margin-left: 10px;
+}
+
+.product-table {
+  width: 100%;
+  border-collapse: collapse;
+  margin-top: 20px;
+}
+
+.product-table th,
+.product-table td {
+  border: 1px solid #ddd;
+  padding: 8px;
+  text-align: left;
+}
+
+.product-table th {
+  background-color: #f4f4f4;
+  font-weight: bold;
+}
+
+.update-button,
+.delete-button {
+  padding: 5px 10px;
+  margin: 0;
+  cursor: pointer;
+}
+
+.update-button {
+  background-color: #007bff;
+  color: white;
+  border: none;
+  padding: 5px 10px;
+  margin-right: 10px; /* Adiciona margem entre os botões */
+  cursor: pointer;
+  border-radius: 5px;
 }
